Update superagent dependency
`superagent` 3.3.1 has a dependency on a vulnerable version of the `mime` package. 

This update bumps the `superagent` dependency to a version which includes a patch to internally update their dependency on the `mime` package.

More information about the `mime` package vulnerability: https://nodesecurity.io/advisories/535

All tests pass after making this update.